async_file_storage/requirements.txt
aiohttp==3.9.5
aiofiles==23.1.0
pyyaml==6.0.1
